requirements.txt
APScheduler==3.6.3
autopep8==1.6.0
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
gTTS==2.2.3
idna==3.3
nodeenv==1.6.0
pycodestyle==2.8.0
pyright==0.0.13
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
